requirements.txt
asyncio==3.4.3
aiogram==3.20.0.post0
dotenv==0.9.9
python-dotenv==1.1.1
requests==2.32.5
eclips-blossom-ai==0.2.2